requirements.txt
pathvalidate ~= 3.2
pyyaml ~= 6.0
tldextract ~= 3.4
yt-dlp >= 2024.5.27